refactor: isSwapRouterContractConfig to isUniquePerPoolContractConfig, supporting NativeTokenGateway
