+++ apps/evm/src/libs/contracts/scripts/generateContractRecords/generateContracts/generateTypes/typesTemplate.hbs
@@ -2,6 +2,6 @@ export type UniqueContractName = {{#each uniques}}| '{{this}}'{{/each}};
 
 export type GenericContractName = {{#each generics}}| '{{this}}'{{/each}};
 
-export type SwapRouterContractName = '{{swapRouter}}';
+export type UniquePerPoolContractName = {{#each uniquePerPool}}| '{{this}}'{{/each}};
 
-export type ContractName = UniqueContractName | GenericContractName | SwapRouterContractName;
+export type ContractName = UniqueContractName | GenericContractName | UniquePerPoolContractName;
