--- apps/evm/src/libs/contracts/utilities/isSwapRouterContractConfig/index.ts
@@ -1,3 +1,0 @@
-import { ContractConfig } from 'libs/contracts/config';
-
-export const isSwapRouterContractConfig = ({ name }: ContractConfig) => name === 'SwapRouter';
