+++ apps/evm/src/libs/contracts/utilities/isUniquePerPoolContractConfig/index.ts
@@ -1,0 +1,6 @@
+import { ContractConfig } from 'libs/contracts/config';
+
+const uniquePerPoolContracts = ['SwapRouter', 'NativeTokenGateway'];
+
+export const isUniquePerPoolContractConfig = ({ name }: ContractConfig) =>
+  uniquePerPoolContracts.includes(name);
